Moved back some functions to middleware
